+++ src/react-components/src/components/02-molecules/menus/Breadcrumbs/Breadcrumbs.stories.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import Breadcrumb from "./Breadcrumb";
+import Breadcrumb from "./Breadcrumbs";
 import { action } from "@storybook/addon-actions";
 
 export default {
@@ -10,7 +10,7 @@ export default {
 };
 
 const shortItems = [
-  { url: "#", text: "Parent" }, 
+  { url: "#", text: "Parent" },
   { url: "#", text: "Home" }
 ];
 
@@ -21,5 +21,10 @@ const longItems = [
   { url: "#", text: "Roast Half And Half Pumpkin Spice Siphon Aroma Ristretto Cinnamon Saucer" }
 ];
 
+const components = [
+  <a role="link" href="#test1"><span>Hello</span></a>, <a role="link" href="#test2">Goodbye</a>
+];
+
 export const shortBreadcrumbs = () => <Breadcrumb breadcrumbs={shortItems}></Breadcrumb>;
 export const longBreadcrumbs = () => <Breadcrumb breadcrumbs={longItems}></Breadcrumb>;
+export const breadcrumbWithComponents = () => <Breadcrumb breadcrumbs={components}></Breadcrumb>;