+++ src/react-components/CHANGELOG.md
@@ -6,6 +6,10 @@ Currently, this repo is in Prerelease.  When it is released, this project will a
 
 ========
 ## Prerelease
+## [0.0.3]
+### Changed
+- Changed `Breadcrumb` to `Breadcrumbs` to match Twig
+- Changed the Breadcrumbs contract to accept components
 
 ## [0.0.2] - 2019-11-13
 ### Added